+++ src/libshadertrap/src/command_create_empty_texture_2d.cc
@@ -22,7 +22,7 @@ namespace shadertrap {
 
 CommandCreateEmptyTexture2D::CommandCreateEmptyTexture2D(
     std::unique_ptr<Token> start_token, std::string result_identifier,
-    uint32_t width, uint32_t height)
+    size_t width, size_t height)
     : Command(std::move(start_token)),
       result_identifier_(std::move(result_identifier)),
       width_(width),
